pyproject.toml
[project]
name = "mema-bricks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["pygame-ce>=2.5.3"]

[dependency-groups]
dev = ["ruff>=0.9.7"]

[tool.ruff]
line-length = 100
target-version = "py312"
